+++ backEnd/app/mvn-start/src/main/java/local/arch/infrastructure/controller/news/NewsAdmin.java
@@ -3,10 +3,11 @@ package local.arch.infrastructure.controller.news;
 import jakarta.inject.Inject;
 import jakarta.json.Json;
 import jakarta.json.JsonArrayBuilder;
-import jakarta.validation.ConstraintViolationException;
+import jakarta.json.JsonObjectBuilder;
 import jakarta.validation.Valid;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.DELETE;
+import jakarta.ws.rs.HeaderParam;
 import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.PUT;
@@ -14,15 +15,28 @@ import jakarta.ws.rs.Path;
 import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.Response;
-
-import local.arch.application.interfaces.news.INewsService;
-import local.arch.domain.entities.News;
+import local.arch.application.interfaces.config.IFileConfig;
+import local.arch.application.interfaces.page.news.INewsService;
+import local.arch.domain.entities.page.News;
 import local.arch.infrastructure.builder.news_annotation.BuiltNews;
+import local.arch.infrastructure.token.ITokenKey;
 
 @Path("/admin/news")
 public class NewsAdmin {
 
     JsonArrayBuilder arrayBuilder = Json.createArrayBuilder();
+
+    private JsonObjectBuilder buildErrorMessage(String msg, Boolean status) {
+        return Json.createObjectBuilder()
+                .add("message", msg)
+                .add("status", status);
+    }
+
+    @Inject
+    IFileConfig fileConfig;
+
+    @Inject
+    ITokenKey tokenKey;
 
     @Inject
     @BuiltNews
@@ -32,12 +46,42 @@ public class NewsAdmin {
     @POST
     @Produces("application/json")
     @Consumes("application/json")
-    public Response addNewsInfo(@Valid News newsInfo) {
+    public Response addNewsInfo(
+            @Valid News newsInfo,
+            @HeaderParam("token") String token) {
+
         try {
+            if (!tokenKey.isTokenValid(token)) {
+                return Response.status(Response.Status.UNAUTHORIZED)
+                        .entity(buildErrorMessage("Недействительный токен", false).build())
+                        .build();
+            }
+
+            if (!tokenKey.isAdmin(token)) {
+                return Response.status(Response.Status.FORBIDDEN)
+                        .entity(buildErrorMessage("Недостаточно прав", false).build())
+                        .build();
+            }
+
+            if (newsInfo.getImage() != null && !newsInfo.getImage().isBlank()) {
+                String imageUrl = fileConfig.saveImageFromBase64(newsInfo.getImage(), "news/1" );
+                newsInfo.setImageUrl(imageUrl);
+            }
+
             newsService.addNews(newsInfo);
-            return Response.ok().build();
-        } catch (ConstraintViolationException e) {
-            throw e;
+
+            return Response.ok(buildErrorMessage("Данные сохранены", true).build())
+                    .build();
+        } catch (NotFoundException e) {
+            return Response
+                    .ok(buildErrorMessage("Произошла ошибка: " + e, false).build())
+                    .status(Response.Status.NOT_FOUND)
+                    .build();
+        } catch (Exception e) {
+            return Response
+                    .ok(buildErrorMessage("Произошла ошибка: " + e, false).build())
+                    .status(Response.Status.INTERNAL_SERVER_ERROR)
+                    .build();
         }
     }
 
@@ -45,43 +89,80 @@ public class NewsAdmin {
     @PUT
     @Produces("application/json")
     @Consumes("application/json")
-    public Response changeNewsInfo(@Valid News newsInfo) {
+    public Response changeNewsInfo(
+            @Valid News newsInfo,
+            @HeaderParam("token") String token) {
 
         // изменить инфу о новости
 
         try {
-            newsService.changeNewsInfo(newsInfo.getNewsID(), newsInfo);
-            return Response.ok().build();
+            if (!tokenKey.isTokenValid(token)) {
+                return Response.status(Response.Status.UNAUTHORIZED)
+                        .entity(buildErrorMessage("Недействительный токен", false).build())
+                        .build();
+            }
 
-        } catch (ConstraintViolationException e) {
-            throw e;
+            if (!tokenKey.isAdmin(token)) {
+                return Response.status(Response.Status.FORBIDDEN)
+                        .entity(buildErrorMessage("Недостаточно прав", false).build())
+                        .build();
+            }
+
+            newsService.changeNewsInfo(newsInfo.getNewsID(), newsInfo);
+
+            return Response.ok(buildErrorMessage("Данные сохранены", true).build())
+                    .build();
+        } catch (NotFoundException e) {
+            return Response
+                    .ok(buildErrorMessage("Произошла ошибка: " + e, false).build())
+                    .status(Response.Status.NOT_FOUND)
+                    .build();
+        } catch (Exception e) {
+            return Response
+                    .ok(buildErrorMessage("Произошла ошибка: " + e, false).build())
+                    .status(Response.Status.INTERNAL_SERVER_ERROR)
+                    .build();
         }
     }
 
     // ?
     @DELETE
     @Path("/{newsID}")
-    public Response deleteNewsAdmin(@PathParam("newsID") Integer newsID) {
-        if (newsID != null) {
-            try {
-                newsService.deleteNews(newsID);
-                return Response
-                        .ok("{\n\"status\": true \n}")
+    public Response deleteNewsAdmin(
+            @PathParam("newsID") Integer newsID,
+            @HeaderParam("token") String token) {
+
+        // удалить новость 
+
+        try {
+            if (!tokenKey.isTokenValid(token)) {
+                return Response.status(Response.Status.UNAUTHORIZED)
+                        .entity(buildErrorMessage("Недействительный токен", false).build())
                         .build();
-
-            } catch (NotFoundException e) {
-                return Response
-                        .status(Response.Status.NOT_FOUND)
-                        .entity("{\"status\": false, \n\"message\": \"Новость не найдена\" \n}").build();
-            } catch (Exception e) {
-                return Response
-                        .status(Response.Status.INTERNAL_SERVER_ERROR)
-                        .entity("{\n\"status\": false, \n\"message\": \"Внутренняя ошибка сервера\" \n}").build();
             }
-        } else {
-            return Response.ok("{\n\"status\": false \n\"message\": \"Неверный ID\" \n}").build();
-        }
 
-        // удалить мероприятие
+            if (!tokenKey.isAdmin(token)) {
+                return Response.status(Response.Status.FORBIDDEN)
+                        .entity(buildErrorMessage("Недостаточно прав", false).build())
+                        .build();
+            }
+            if (newsID != null) {
+                newsService.deleteNews(newsID);
+
+                return Response.ok(buildErrorMessage("Урок удалён", true).build()).build();
+            } else {
+                return Response.ok(buildErrorMessage("Неверный ID", false).build()).build();
+            }
+        } catch (NotFoundException e) {
+            return Response
+                    .ok(buildErrorMessage("Произошла ошибка: " + e, false).build())
+                    .status(Response.Status.NOT_FOUND)
+                    .build();
+        } catch (Exception e) {
+            return Response
+                    .ok(buildErrorMessage("Произошла ошибка: " + e, false).build())
+                    .status(Response.Status.INTERNAL_SERVER_ERROR)
+                    .build();
+        }
     }
 }
